requirements/base.txt
# Base
pytz==2018.7
python-slugify==1.2.6
Pillow>=6.2.0
#psycopg2==2.7.4 --no-binary psycopg2

# Django
django==2.2.10

# Environment
django-environ==0.4.5

# Passwords security
argon2-cffi==18.3.0

# Static files
whitenoise==4.1.2

# Django CK Editor
django-ckeditor

